+++ tiles/src/components/commission/form_steps/CommissionSummary.js
@@ -1,90 +1,100 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Form from "react-bootstrap/Form";
-import Table from "react-bootstrap/Table";
-import { displayZloty } from "../../../utils";
-import Row from "react-bootstrap/Row";
-import Col from "react-bootstrap/Col";
-import { COMMISSION_STATUS, VEHICLE } from "../choices";
 
-const CommissionSummary = ({ currentStep, commission }) => {
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faExclamationCircle } from "@fortawesome/free-solid-svg-icons";
+
+import Table from "react-bootstrap/Table";
+import Card from "react-bootstrap/Card";
+import Accordion from "react-bootstrap/Accordion";
+import Alert from "react-bootstrap/Alert";
+
+import VehicleCard from "../VehicleCard";
+import ContractorCard from "../../invoicing/ContractorCard";
+import ComponentCard from "../ComponentCard";
+import { displayZloty } from "../../../utils";
+import StatusButtons from "./StatusButtons";
+
+const CommissionSummary = ({ currentStep, commission, onChange }) => {
     if (currentStep !== 8) return null;
 
     return (
-        <Form.Group>
-            <Row>
-                <Col>
-                    <Table bordered className="table-row-headers">
-                        <tbody>
-                            <tr>
-                                <td>Typ zlecenia</td>
-                                <td>{commission.commission_type === VEHICLE ? "Pojazd" : "Podzespół"}</td>
-                            </tr>
-                            <tr>
-                                <td>Status zlecenia</td>
-                                <td>{COMMISSION_STATUS.filter((x) => x.key === commission.status)[0].label}</td>
-                            </tr>
-                            <tr>
-                                <td>{commission.commission_type === VEHICLE ? "Pojazd" : "Podzespół"}</td>
-                                <td>{commission.vc_name}</td>
-                            </tr>
-                            <tr>
-                                <td>Kontrahent</td>
-                                <td>{commission.contractorLabel}</td>
-                            </tr>
-                            <tr>
-                                <td>Opis</td>
-                                <td>{commission.description}</td>
-                            </tr>
-                            <tr>
-                                <td>Data przyjęcia</td>
-                                <td>{commission.start_date}</td>
-                            </tr>
-                            <tr>
-                                <td>Data zakończenia</td>
-                                <td>{commission.end_date ? commission.end_date : ""}</td>
-                            </tr>
-                        </tbody>
-                    </Table>
-                </Col>
+        <>
+            <div className="d-flex justify-content-between align-items-center flex-wrap">
+                <div className="text-left">
+                    <header className="font-weight-bold">Data przyjęcia</header>
+                    <div>{commission.start_date}</div>
+                </div>
+                <StatusButtons commission={commission} onChange={onChange} size="xl" />
+                <div className="text-right">
+                    <header className="font-weight-bold">Data zamknięcia</header>
+                    <div>{commission.end_date ? commission.end_date : "—"}</div>
+                </div>
+            </div>
 
-                {commission.items.length > 0 ? (
-                    <Col>
-                        <Table striped bordered hover responsive>
-                            <thead>
-                                <tr>
-                                    <th className="th-ware">Usługa/Towar</th>
-                                    <th className="th-price">Cena</th>
-                                    <th className="th-quantity">Ilość</th>
-                                    <th className="th-sum">Wartość</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                {commission.items.map((item, index) => (
-                                    <tr key={item.id || `new_${index}`}>
-                                        <td>
-                                            <div>{item.name}</div>
-                                            {item.description ? <div className="mt-1">{item.description}</div> : null}
-                                        </td>
-                                        <td>{displayZloty(item.price)}</td>
-                                        <td>{item.quantity}</td>
-                                        <td>{displayZloty(item.price * item.quantity)}</td>
-                                    </tr>
-                                ))}
-                            </tbody>
-                            <tfoot>
-                                <tr>
-                                    <td colSpan={3}>Razem:</td>
-                                    <td>
-                                        {displayZloty(commission.items.reduce((a, b) => a + b.price * b.quantity, 0))}
-                                    </td>
-                                </tr>
-                            </tfoot>
-                        </Table>
-                    </Col>
+            {commission.description ? (
+                <Alert className="mt-4 detail-description-alert" variant="primary">
+                    <div className="d-flex align-items-center">
+                        <FontAwesomeIcon icon={faExclamationCircle} size="3x" className="mr-2" />
+                        {commission.description}
+                    </div>
+                </Alert>
+            ) : null}
+
+            <Accordion className="mt-4" defaultActiveKey="items">
+                {commission.vehicle ? <VehicleCard id={commission.vehicle} bg="light" border="primary" /> : null}
+                {commission.component ? <ComponentCard id={commission.component} bg="light" border="primary" /> : null}
+                {commission.contractor ? (
+                    <ContractorCard id={commission.contractor} bg="light" border="primary" />
                 ) : null}
-            </Row>
-        </Form.Group>
+                {commission.items.length > 0 ? (
+                    <Card bg="light" border="primary">
+                        <Accordion.Toggle as={Card.Header} eventKey="items">
+                            Pozycje
+                        </Accordion.Toggle>
+                        <Accordion.Collapse eventKey="items">
+                            <Card.Body>
+                                <Table striped bordered hover responsive>
+                                    <thead>
+                                        <tr>
+                                            <th className="th-ware">Usługa/Towar</th>
+                                            <th className="th-price">Cena</th>
+                                            <th className="th-quantity">Ilość</th>
+                                            <th className="th-sum">Wartość</th>
+                                        </tr>
+                                    </thead>
+                                    <tbody>
+                                        {commission.items.map((item, index) => (
+                                            <tr key={item.id || `new_${index}`}>
+                                                <td>
+                                                    <div>{item.name}</div>
+                                                    {item.description ? (
+                                                        <div className="mt-1">{item.description}</div>
+                                                    ) : null}
+                                                </td>
+                                                <td>{displayZloty(item.price)}</td>
+                                                <td>{item.quantity}</td>
+                                                <td>{displayZloty(item.price * item.quantity)}</td>
+                                            </tr>
+                                        ))}
+                                    </tbody>
+                                    <tfoot>
+                                        <tr>
+                                            <td colSpan={3}>Razem:</td>
+                                            <td>
+                                                {displayZloty(
+                                                    commission.items.reduce((a, b) => a + b.price * b.quantity, 0)
+                                                )}
+                                            </td>
+                                        </tr>
+                                    </tfoot>
+                                </Table>
+                            </Card.Body>
+                        </Accordion.Collapse>
+                    </Card>
+                ) : null}
+            </Accordion>
+        </>
     );
 };